+++ calculator/src/main/java/edu/nextstep/camp/calculator/CalculatorHistoryAdapter.kt
@@ -4,13 +4,12 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
 import edu.nextstep.camp.calculator.databinding.ItemResultBinding
 import edu.nextstep.camp.calculator.domain.CalculateHistory
 
-class CalculatorHistoryAdapter : ListAdapter<CalculateHistory, CalculatorResultViewHolder>(diffUtil) {
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CalculatorResultViewHolder {
-        return CalculatorResultViewHolder(
+class CalculatorHistoryAdapter : ListAdapter<CalculateHistory, CalculatorHistoryViewHolder>(diffUtil) {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CalculatorHistoryViewHolder {
+        return CalculatorHistoryViewHolder(
             ItemResultBinding.inflate(
                 LayoutInflater.from(parent.context),
                 parent,
@@ -19,7 +18,7 @@ class CalculatorHistoryAdapter : ListAdapter<CalculateHistory, CalculatorResultV
         )
     }
 
-    override fun onBindViewHolder(holder: CalculatorResultViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: CalculatorHistoryViewHolder, position: Int) {
         holder.onBind(getItem(position))
     }
 
@@ -37,14 +36,5 @@ class CalculatorHistoryAdapter : ListAdapter<CalculateHistory, CalculatorResultV
                 newItem: CalculateHistory,
             ): Boolean = oldItem == newItem
         }
-    }
-}
-
-class CalculatorResultViewHolder(private val binding: ItemResultBinding) :
-    RecyclerView.ViewHolder(binding.root) {
-    fun onBind(calculateResult: CalculateHistory) {
-        binding.tvExpression.text = calculateResult.expression.toString()
-        binding.tvResult.text =
-            binding.root.resources.getString(R.string.calculate_result, calculateResult.result)
     }
 }
